+++ app/gui/gui.js
@@ -192,8 +192,14 @@ async function genPageHeader ( pgHeader, req, page ) {
         let ident = ''
         let deepth = (scope.match(/\//g) || []).length
         for ( let i = 0; i < deepth; i++ ) { ident += '&nbsp;&nbsp;'}
-        menuItems.push({ html: ident + '<a href="setscope?id='+scope+'&layout='+page+'">'+scopeTbl[ scope ].name+'</a>' })
+        menuItems.push({ html: ident + '<a href="setscope?id='+scope+'&layout='+page+'">'+scopeTbl[ scope ].name+'</a>', id: scope })
       }
+
+      menuItems.sort( ( a, b ) => {
+        if ( a.id > b.id ) { return 1 }
+        return -1
+      })
+
       let actScope = await userDta.getSelScopeName( user.userId )
       pgHeader.modules.push({ 
         id    : "ScopeSel",